pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-ml-learning-journey"
version = "0.1.0"
description = "AI/ML Learning Journey with FastAPI examples"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "scalar-fastapi>=1.5.0",
    "sqlalchemy>=1.4.36,<2.0.0",
    "psycopg2-binary>=2.9.11",
    "apache-airflow>=2.10.0,<3.0.0",
    "gensim==4.3.2",
    "scipy<1.13.0",
    "matplotlib",
    "seaborn",
    "numpy",
    "python-dotenv"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["concepts*", "projects*"]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",  # for testing FastAPI
]
